Add local blocks to the FFT Block+Cyclic code.
To do this, I had to add "dont disable remote value forwarding" pragmas to a
few functions in ChapelArray.chpl. I added a new help function to them, used
the pragma on the help function, and do the sync accesses in it. I did this
minimally, only for the functions that need it for this case.

Changed the name of Block distribution and domain variables from ProblemDist
and ProblemDom to BlkDist and BlkDom.

Pushed the logic mapping the current phase to the iteration space into the
genDFTStrideSpan iterator, which Brad suggested.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16347 3a8e244f-b0f2-452b-bcba-4c88e055c3ca