Reworked the use of function pointers during the transformation of recursive iterators into recursive functions to use instead an integer into the function pointer table chpl_ftable.  This removes the assumption that function pointers will be the same on different locales.  This change also eliminates the need for the FnType class although I haven't removed it as part of this change.
Details
-------

Added a new primitive PRIM_FTABLE_CALL to deal with indirect function
calls via the generated table of function pointers chpl_ftable.
Changed how this table is handled so that functions can be added to
this table throughout compilation.  Modified prune() so as not to
delete functions that are only called via this table.

Changed the caching of iterator functions used to terminate the
transformation from recursive iterators to recursive functions so that
the loop body functions are cached alongside the iterator functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16348 3a8e244f-b0f2-452b-bcba-4c88e055c3ca