Filing a bug that John Lewis ran into today
It appears that when we have a class declaration within a generic
function we don't create a defaultTypeConstructor for it, resulting
in a seg fault.  I've captured the issue in the small test
locClassSegFault and am also filing a version that doesn't use a
generic function to show that it works.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16351 3a8e244f-b0f2-452b-bcba-4c88e055c3ca