This change makes the PVM group name in the comm layer unique across runs.
This change makes the PVM group name in the comm layer unique across runs. Before this change, the job name was hardcoded to be "job." Now, the job name is a global (static) variable which, in chpl_comm_init, gets the name job12345 representing the pid of the launcher. Should you be starting the _real directly with gdb, you now need another argument (typeProblem --localesPerRealm='1 1' 2 12345) representing an ID (which doesn't have to be the PID).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16370 3a8e244f-b0f2-452b-bcba-4c88e055c3ca