This change moves PVM comm layer code around.
This change moves the gates dictating when to stop doing barriers (which dictates the flow of how the PVM code exits) and when to start polling from the rollcall (where it isn't really appropriate other than the fact it used to be the last place I could catch it before entering main) to a special case of chpl_comm_barrier. In main.c, the last thing that's done before the root node enters chpl_main is a barrier ("barrier before main"). When chpl_comm_barrier catches that, the code now tells the barrier code to stop working and the poller to start.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16371 3a8e244f-b0f2-452b-bcba-4c88e055c3ca