Optimized recursive iterator to function transform.
Inserted code to do a deep copy (recursive) of the argument bundle
class before calling the iterator function if the iterator contains
any on-statements.  This improves locality by ensuring that the
argument bundle class is allocated locally.  We could optimize this in
certain cases by not doing the copy if the class is already local.

Made remote value forwarding more robust with respect to reference
fields of argument bundle classes by ignoring assignments of the field
from one instance of the class into the same field of another instance
(or the same instance if that happens to become the case) of the
class.

Inserted code to free the argument bundle class.  The deep copy is not
yet freed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16375 3a8e244f-b0f2-452b-bcba-4c88e055c3ca