This change makes the PVM comm layer exit more guarenteed.
This change makes the PVM comm layer exit dependent on the polling thread having exited. Before this change, it was possible for the polling thread to still be doing something before exiting leading to PVM errors on exit (race). With the reverse start-polling okaypoll mechanism in place, this race is gone.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16376 3a8e244f-b0f2-452b-bcba-4c88e055c3ca