This change ensures a PVM group order in the comm layer.
This change ensures a PVM group order in the comm layer. It also adds another argument in the pvm_spawn: an order number (therefore, when using gdb, you will now need the total number of locales, an ID number you can make up on the fly but should be different between runs, and an order of the nodes). The comm layer uses that order number to wait until there are that many elements in the group before joining. (If there aren't any elements in the group, pvm_gsize returns PvmNoGroup, so the 0th node has to be its own case.)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16378 3a8e244f-b0f2-452b-bcba-4c88e055c3ca