Fixing an overflow problem in range slicing
When slicing ranges, we computed the stride of the result by
multiplying the two original strides and dividing by the gcd:

  s1 * s2 / gcd

Doing these operations L->R is likely to result in overflow
if the strides are sizeable, so this change does one of
the divisions first and then the multiplication.  This
permits our FFT code to run at full problem size on one
(and presumably more) processors.

Checking in a test to recreate the approximate behavior I
was seeing and lock in this fix.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16382 3a8e244f-b0f2-452b-bcba-4c88e055c3ca