Added "fast on" optimization.
This commit adds the "fast on" optimization.  If the body of an on
statement is found to be "short and safe" (see below), the wrapper
function for the on statement is marked as a candidate for "fast on".
The implementation for a given communication layer can define
chpl_comm_fork_fast() to take advantage of this information.  For
example, the gasnet implementation will run the wrapper directly in
the remote handler, rather than creating a new task.  Currently, only
the gasnet implementation takes advantage of this optimization.

Currently "short and safe" is defined as follows:

- no communication
- no sync/single access
- function call depth less than 20 (adjustable by compile time flag)

Probably there should be some tuning with the the call depth, as well
as add a limit on the tall instruction count.

This optimization has been shown to improve RA performance by close to
2x.


COMPILER CHANGES

- Implemented markFastOn() pass in compiler/passes/parallel.cpp
- Added new AST flag, FLAG_FAST_ON
- Added codegen path for chpl_comm_fork_fast() if FLAG_FAST_ON is set
- Added new compiler flags:

      --[no-]fast-on                     Enable [disable] fast remote on
      --fast-on-limit <limit>            Limit recursion depth of fast on
                                         optimization search


RUNTIME CHANGES

- Added chpl_comm_fork_fast() to communication layer interface
- Implemented chpl_comm_fork_fast() for gasnet
- For other comm layers, default to chpl_comm_fork()


TEST CHANGES

- release/examples/hpcc/ra.chpl: added "local" block to main loop to
  enable fast on optimization

- compflags/bradc/help/userhelp.good: Fixed compiler help output



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16383 3a8e244f-b0f2-452b-bcba-4c88e055c3ca