Updating FFT to reflect Block + Cyclic distributions
I wanted to get a copy of FFT into the release directory for broader
testing tonight.  This is a version that started from
test/studies/hpcc/FFT/diten/fft.chpl, but then I edited it to update
comments, improve self-documenting code, remove dead code, etc.
Comments welcome (particularly from David).

In writing this, I realized that we're still using the forall + on
clause approach in the inner parallel loop.  I like David's proposal
to use the sliced domain but didn't tackle that here.  David, any
reason not to switch over to it?  (note that this means my memory
tests this weekend didn't use it either...).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16393 3a8e244f-b0f2-452b-bcba-4c88e055c3ca