Fix to BlockDist
In our implementation of BlockDist, we have had a longstanding bug in
which the locale->index set mapping ("Which indices do I own as locale
3?  [mylo..myhi]?  Great!") has not quite been the inverse of our
index->locale mapping (e.g., asking "Who owns myhi?" would not return
locale 3 but rather locale 4.).  The chunks that the locales were
grabbing were contiguous and disjoint, so we weren't missing any
elements, but if we did a random access, we could end up accessing
outside of a node's local portion of the array element.  Because this
only shows up at extreme problem sizes and numbers of locales (I
suspect sizes where the product of the problem size and # locale
exceed the integer capabilities of a 64-bit floating point number?)
and because we don't usually have bounds checking on at these sizes,
we haven't been catching this.  Running RA on 32 locales with bounds
checking on did catch it.

The general challenge here is that we want our index->locale mapping
to be a simple integer division which results in a floor.  This
suggests that in computing the locale->index set mapping we want to
take a ceiling.

In the old scheme, we used floating point values to store divisions
and help compute the ceiling by transcribing the code we used in ZPL.
Either I mistranscribed the ZPL or ZPL had bugs that we never ran into
as a result of not running large enough problems (which is very
conceivable because while we ran large problems with ZPL, they were
higher-dimensional problems which are less likely to run into these
overflow values because each dimension is much smaller than what we
see with RA or Stream for example).

In the new scheme, I use a trick for taking the ceiling of x/y where x
and y are integers of adding y-1 to x.  I believe that this will still
have problems at extreme scale -- in particular if #locales *
problemsize in a dimension > 2**63 or 2**64.  Back of the envelope
calculations suggest we're still a ways away from that.  And it could
be that integer wraparound would cause this to work anyway.  Requires
more thought, but for now the previous bug is fixed.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16397 3a8e244f-b0f2-452b-bcba-4c88e055c3ca