This testing change is temporary to debug pvm_addhosts failures.
This testing change is temporary to debug pvm_addhosts failures. As of this weekend, we've been seeing large negative numbers returned in the infos array in the pvm_addhosts call. This change checks to see if this is a result of uninitialized values (thus settling a side bet between Brad and I about pvm_addhost functionality) and builds an error message as it goes. This change will be backed out after tonight.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16403 3a8e244f-b0f2-452b-bcba-4c88e055c3ca