Removing the HPCC documents from the release
On the grounds that the tutorial is out of date, that we won't
have our 2009 report ready in time, and that this material is
arguably better suited for the web where we can keep it up-to-date
(or let it drift out of date) more easily.

Updated the READMEs that referenced these documents.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16405 3a8e244f-b0f2-452b-bcba-4c88e055c3ca