This change makes the PVM launcher changes more permanent.
This change makes the PVM launcher changes from yesterday more permanent. Before adding a new host, we initialize each value in the infos array (status per node) to -33333. Then, if any nodes aren't successful in a pvm_addhosts, we display the error message if verbosity is on. (If verbosity is not on, we suggest turning it on.)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16416 3a8e244f-b0f2-452b-bcba-4c88e055c3ca