Minor fixes to ARMCI communication layer
Took a quick look at the ARMCI communication layer based on
a conversation at SC09 with Abhinav Vishnu who was interested
in getting the ARMCI version working for once and for all.
I fixed some low-hanging fruit that had changed in the interface,
but one major fix is still required in order to get it to compile:
we changed the chpl_comm_fork interface to take an enumeration 
rather than a function pointer in order to deal with the case
where a function pointer may not be symmetric across processors,
and ARMCI was using chpl_comm_fork to implement its broadcasts
but was not kept up-to-date with this change.  Will tackle that
as a separate step.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16435 3a8e244f-b0f2-452b-bcba-4c88e055c3ca