Refactored the implementation of homogeneous tuples (aka star tuples) to allow us to code generate them as static C arrays instead of structs.  For large tuples (many components), this is a big win.
Details
-------

Chapel-level changes:

Changed the Chapel and C names of star tuple types to use a standard
concise form.  For example,

  Before this change:  Chapel:  (int, int, int)
                       C     :  _tuple_3_int32_t_int32_t_int32_t

  After this change :  Chapel:  3*int
                       C     :  _tuple_3_star_int32_t

Updated several tests to use the new Chapel type names.

Runtime changes:

Added a macro CHPL_ASSIGN_SVEC that takes two arguments and does a
memcpy from the second argument to the first argument.  This is used
to implement C-level assignment for star tuples since C does not
support assignment of static arrays.

Added SVEC variants of _ARRAY_SET and all of the CHPL_COMM_WIDE macros
that required it.  Since these communication macros short-circuit and
do direct assignment when things are local, they need a second version
to do the memcpy instead of the direct assignment if the base type is
a star tuple.  Updated code generation to use these new macros when
necessary.  Cleaned and orthogonalized in the process, though there is
now an extra case to consider.  I have more simplification ideas that
I plan to implement.

Module changes:

Removed some commented-out code from the ChapelTuple module and
reworked the tuple indexing function that takes a non-parameter
integer so that it gives a better error message if it is called on
non-star tuples and so that it uses the "get tuple member" primitive
(rather than a conditional statement in an unrolled parameter for
loop!).

Compiler changes:

Added a new flag FLAG_STAR_TUPLE that is applied to tuples
after they are instantiated if all of the component types are
identical (if they are star tuples).

Added a new primitive PRIM_IS_STAR_TUPLE_TYPE that returns true if
the type is a star tuple.  This is used to issue a good compiler error
message in the tuple indexing function.

Add new primitives PRIM_GET_SVEC_MEMBER, PRIM_GET_SVEC_MEMBER_VALUE,
and PRIM_SET_SVEC_MEMBER for accessing and setting the components of a
star tuple (static C vector).  These correspond to the class and
record primitives PRIM_GET_MEMBER, PRIM_GET_MEMBER_VALUE, and
PRIM_SET_MEMBER.  However, instead of taking a field as the second
argument, they take an integral index.

Added a new pass returnStarTuplesByReferenceArguments that changes all
functions that return a star tuple into functions that take a star
tuple by reference and return via this reference argument.  This works
around a limitation in C in which static arrays cannot be returned by
functions.  This pass is run late in compilation, just before wide
references are inserted.  This pass also changes all record access/set
primitives to the new star tuple access/set primitives for star
tuples.  Prior to this point, star tuples are sometimes accessed with
these record primitives.  Indeed, the star tuples are not simplified
in all respects until code generation, meaning that among other
things, they do have a field associated with each component in the
internal representation.  This allows normal optimizations, like
scalar replacement, to continue to work.  To make our implementation
better in the presence of very large star tuples, we'll want to
address this issue.

Added a special case to code generate star tuples as static C arrays
instead of structs.  This is tested at the beginning of
ClassType::codegenDef.

Extended a few transforms in insertWideReferences that simplify cases
involving the SET_MEMBER primitive to also apply to the
SET_SVEC_MEMBER primitive.  These involve string literals and nil,
both of which are widened before use in this primitive.

Extended cullOverReferences to cull references associated with the new
star tuple access and set primitives similarly to the class access and
set primitives.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16448 3a8e244f-b0f2-452b-bcba-4c88e055c3ca