Refactored CallExpr::codegen() to make it more concise, readable, and manageable.  It got especially ugly during my introduction of static C arrays to handle star tuples.
Introduced a new function

  gen(FILE* outfile, const char* format, ...)

that is similar to fprintf except its format specifier codes are
different.  It can generate any BaseAST* passed to it using the format
specifier code '%A'.  For example, we can thus change

  case PRIM_SET_REF:
    fprintf(outfile, "&(");
    get(1)->codegen(outfile);
    fprintf(outfile, ")");

to

  case PRIM_SET_REF:
    gen(outfile, "&(%A)", get(1));

The only other format specifier code recognized now allows us to
generate '%' itself by writing '%%'.

There is more to update, but this is a good start and a good place to
checkpoint.  This eliminates 323 lines of code which is welcome
especially since my commit that introduced the need to call memcpy and
specialize the star tuple case added 201 lines of code in this file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16453 3a8e244f-b0f2-452b-bcba-4c88e055c3ca