Added the following methods to help get types, reference types, and wide reference types:
  Type* BaseAST::getValType()

    This method returns the value type associated with a reference or
    wide reference type, and otherwise returns NULL.  It replaces
    Type::getValueType() and has the advantage that it can be called
    on expressions and symbols as well as types.  It applies to the
    type of the expression or symbol as returned by typeInfo().  This
    advantage makes for simpler code and I updated uses in the current
    code base.  Here are several examples of these changes:

      Before: field->type->getValueType()
      Now   : field->getValType()

      Before: call->get(1)->typeInfo()->getValueType()
      Now   : call->get(1)->getValType()

  Type* BaseAST::getRefType()

    This method returns the reference type associated with any type
    that is not a reference type or wide reference type, and otherwise
    returns NULL.  It replaces the method Type::getReferenceType()
    (which was unused) and has the same advantages as described above.

  Type* BaseAST::getWideRefType()

    This method returns the wide reference type associated with any
    type or reference type, and otherwise returns NULL.

Moving forward, I think it would be advantageous for getValType() to
return the type of 'this' if it is a value type, for getRefType() to
return the type of 'this' if it is a reference type, and for
getWideRefType() to return the type of 'this' if it is a wide
reference type.  This would allow us to simplify the common idiom

  Type* vt = <expression>->typeInfo();
  if (vt->symbol->hasFlag(FLAG_REF))
    vt = vt->getValType();

by changing it to

  Type* vt = <expression>->getValType();

And this would have the advantage of applying to wide reference types
as well.  This common idiom is not more often written as

  Type* vt = <expression>->typeInfo();
  if (vt->symbol->hasFlag(FLAG_REF) || vt->symbol->hasFlag(FLAG_WIDE_REF))
    vt = vt->getValType();

because it is used in passes that are run before insertWideReferences.

It also dawned on me that we can simplify much code like

  <type>->symbol->hasFlag(...)

by implementing hasFlag(...) on types.  This minor cleanup would
result in cleaner lines of code though this complexity can also be
addressed by replacing uses of Type* with TypeSymbol* and more
consistently using TypeSymbol*.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16454 3a8e244f-b0f2-452b-bcba-4c88e055c3ca