Improved the getValType(), getRefType() and getWideRefType() methods by making them more forgiving:
  Changed getValType() to return the type of 'this' if it is neither a
  reference type nor a wide reference type (rather than NULL).

  Changed getRefType() to return the type of 'this' if it is a
  reference type (rather than NULL).

  Changed getWideRefType() to return the type of 'this' if it is a
  wide reference type (rather than NULL).

Simplified the code base in two places where we relied on these
methods returning NULL in these cases.  In both cases, the code is now
clearer.  In the first case, we relied on the return value of NULL to
distinguish between wide references and wide classes.  It is better to
distinguish between these cases based on what they are.  In the second
case, we relied on the return value of NULL in determining what types
we can prune.  I cleaned this code up by removing hidden assumptions
and recommented.

Simplified many other places in the code base as suggested in r16454.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16457 3a8e244f-b0f2-452b-bcba-4c88e055c3ca