Refactored code to handle dynamic dispatch to use conditional statements rather than conditional expressions (as nested functions) resulting in substantially better scaling of inlining (compilation time with respect to program size).  Even so, the code associated with dynamic dispatch seems to be swamping compilation time and I plan to look into how much effort would be required to implement a virtual method table.  I also have a question about the reallocate method which is dynamically dispatched and generic.  This is not allowed in C++, I think, and I'm a little concerned that we may generate far too much code as a result.  That said, it looks like the size of our generated code/intermediate representation scales linearly with program size.
Reworked the CondStmt constructor to avoid inserting an extra
BlockStmt in simple cases.

Added a compiler scaling test "test_many_arrays_of_records" that
contains a configuration parameter "p" used to control the number of
distinct array types that are created.  I have this checked in with
p=32, but it is meant to be increased to p=192 in order to test that
there is no instantiation limit on functions and types in the internal
modules.  It should be increased when compilation takes a more
reasonable amount of time.  Here are some compilation timings as
reported via --print-passes:

                     before this commit   after this commit
                      p=64     p=128       p=64     p=128
resolve                14.931   38.271      14.673   37.448
flattenFunctions        2.244   12.233       1.376    4.846
inlineFunctions         5.613   21.901       1.643    3.190
refPropagation          9.628   42.729       9.739   42.796
copyPropagation        19.156   80.415      18.672   77.186
removeWrapRecords       4.211   19.617       4.027   19.571
deadCodeElimination    14.901   60.240      15.994   63.750
makeBinary             13.775   26.801      13.763   26.538
total time            110.143  360.783     105.466  340.890

As the program size doubles from 64 to 128, we would like for the
compilation time to increase by at most a factor of two.  This commit
fixes the scaling of inlining by changing the representation used to
implement dynamic dispatching.  It also improves flattenFunctions and
copyPropagation, but these do not scale.  Performance of dead code
elimination ebbs a bit.  All of the other listed passes show major
scaling issues in that the time more than doubles.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16470 3a8e244f-b0f2-452b-bcba-4c88e055c3ca