Reworked dynamic dispatch code to use a virtual method table instead of a sequence of conditional statements.
  For certain Chapel codes, the virtual method table significantly
  decreases the number of lines of generated C code.  For example, for
  the compiler scaling test "test_many_arrays_of_records," the number
  of lines of C code generated for the module ChapelDistribution was
  10665 and is now 341.  These results are not typical.  Total lines
  of generated code dropped from 413750 to 392318.

Details
=======

Flag Changes:

Added a new flag --conditional-dynamic-dispatch-limit <limit> that
takes a limit and generates a sequence of conditional statements to
implement dynamic dispatch if the number of candidate methods is less
than or equal to this limit.  According to the literature and online
resources, this is an important compiler optimization since it enables
inlining, however it is rarely implemented.  Since we already had this
implemented, it seemed like a good idea to keep it and add a flag.
The default limit is currently set to zero, but we can change this to
4 once the virtual method table mechanism is tested via the regression
testing system for a bit.  Added help message and man page entry.

Alphabetized --optimize-on-clauses and --optimize-on-clause-limit.

Compiler Changes:

Added a number of maps that are built during function resolution and
which store the virtual method table.  The map virtualMethodTable maps
types to vectors of function symbols and defines the virtual method
table.  The map virtualMethodMap maps functions to integral indices
that represent their offset into the vector of function symbols in the
virtualMethodTable map.  The map virtualRootsMap maps functions to
vectors of function symbols that are roots of any dynamic dispatch
tree involving the source function.  (Until multiple inheritance is
supported, this vector of function symbols will contain exactly one
function symbol.)  The map virtualChildrenMap maps functions to
vectors of function symbols that are children (directly or
transitively) and thus candidates for dynamic dispatch should the
source function be called.  This map was the only one computed before
this commit; it was (and still is) used to generate the sequence of
conditional statements that implements the dynamic dispatch.

Added FLAG_VIRTUAL to mark functions involved in a dynamic dispatch,
including both roots and children in the override tree.

Alphabetized FLAG_INVISIBLE_FN.

Renamed PRIM_GETCID to PRIM_TESTCID.  This primitive tested whether
the class id of an object equals a specified class id.

Added a new primitive PRIM_GETCID that returns the class id of an
object.  This is then used to index into the virtual method table in
order to call a virtual method via dynamic dispatch.

Added new primitive PRIM_GETCID to primitives that need to be nil
checked.  PRIM_TESTCID is one of these primitives as well.

Added a new primitive PRIM_VMT_CALL that takes a function (a root in
the virtual method table), a class id, and an arbitrary number of
function arguments passed to the indirectly called virtual function.

Added code to generate two different virtual method tables, one for
chapel_code.chpl in the runtime and one for an actual Chapel program.
The former is called chpl_rt_vmtable and the latter is called
chpl_vmtable.

Disabled pruning of virtual functions.

Added all virtual functions to the call graph for a virtual function
call.  This changes the semantics of the call graph since a CallExpr
in the calledBy vector may now not be a call of a function symbol but
rather a call of the virtual method call primitive.

Extended function actual_to_formal to work with PRIM_VMT_CALL by
returning a formal in the root function.

Changed insert line numbers to insert line numbers and filenames into
all virtual function parents, children, and siblings when adding line
numbers and filenames to any virtual functions.

Changed inlining so that it no longer assumes that the calledBy vector
does not contain a primitive (since it now can contain a
PRIM_VMT_CALL).  Changed inlining so that virtual functions that have
the inline pragma are not removed.

Runtime changes:

Added a new macro to implement calls into the virtual method table.
Changed the name of the macro for thea TESTCID primitive and added a
new macro for the new GETCID primitive.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16480 3a8e244f-b0f2-452b-bcba-4c88e055c3ca