This commit measurably improves compilation performance.  It rearranges calls to deadVariableElimination, deadExpressionElimination, and deadCodeElimination, makes deadVariableElimination non-iterative, and improves the asymptotic performance of part of copy propagation from quadratic to linear.
Details
=======

Exposed collectSymbolSetSymExprVec which was previously used
internally by buildDefUseMaps.  This allows places where we are going
to build defMaps and useMaps to build the set of Symbols and vector of
SymExprs so that these structures can be used without computing them
twice.

Moved deadCodeElimination before removeWrapRecords so that it
immediately follows copyPropagation, which in turn immediately follows
refPropagation, thus making it natural to remove calls to
deadVariableElimination and deadExpressionElimination from
copyPropagation and refPropagation.  After some experimentation, I
left these calls in copyPropagation between the local and global
phases as this improves compilation time.  The generated code is
unaffected.

Changed the calls to deadVariableElimination and
deadExpressionElimination into calls to deadCodeElimination in
lowerIterators and gpuFlattenArgs.  The function deadCodeElimination
now calls deadVariableElimination and deadExpressionElimination right
before returning (since this was always done at every call site).

Made deadVariableElimination non-iterative.  This iterative algorithm
was not a major source of overhead, but it doesn't need to be
iterative at all if it follows deadCodeElimination.

The above change with no call to deadVariableElimination between the
local and global phases of copy propagation improved compilation time
on the compiler scaling test "test_many_arrays_of_records" but
devastated the compilation time of test/users/franzf/v0/chpl/main.
The time spent in copy propagation increased from about 2.5 seconds to
over 60 seconds.  Looking into this issue led to the following change
to copy propagation:

Reworked the code in copyPropagation in which the KILL sets are
created for each basic block.  For each basic block, we iterate over
the statements.  Each time we found code that would invalidate some
available copies, we would iterate over the KILL set (which has all
available copies).  Now each time we find code that invalidates some
available copies, we add this to a set and a vector.  After we collect
them all, we iterate over the KILL set and look at the set of
invalidated symbols.  This changes the complexity from

      NS * NC   to   NS + NC

where NS is the number of statements and NC is the number of available
copies.  Since NC is on the order of NS, we drop from quadratic to
linear asymptotic performance.

This commit improves compilation speed on the compiler scaling test
"test_many_arrays_of_records" by 24.7% with p=128:

                     before this commit   after this commit
                      p=64     p=128       p=64     p=128
resolve                14.844   38.409      14.988   38.272
flattenFunctions        1.118    5.633       0.979    5.718
inlineFunctions         1.629    3.160       1.644    3.114
refPropagation         10.013   41.489       3.393   15.191
copyPropagation        17.627   68.543       9.627   41.159
deadCodeElimination    14.908   61.617      15.665   66.757
removeWrapRecords       3.694   18.947       0.478    1.744
makeBinary             13.323   26.245      13.354   25.368
total time            106.291  339.870      86.957  255.786

The times reported for before this commit are close to those reported
after commit r16470.

This commit also improves compilation speed on most other tests
including test/users/franzf/v0/chpl/main by 5.7% and
test/arrays/deitz/jacobi.chpl by 6.7%.

The generated code is mostly unaffected.  Due to moving
removeWrapRecords after deadCodeElimination, a nominal number of
lines, mainly in ChapelArray.c, are not eliminated.  The
removeWrapRecords optimization, which is not scaling, should be made
more robust.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16486 3a8e244f-b0f2-452b-bcba-4c88e055c3ca