Optimized two parts of reachingDefinitionsAnalysis resulting in an approximately 1% compilation speedup on the compiler scaling test test_many_arrays_of_records.
Details
=======

Optimized freeDefUseChains which explicitly deletes the heap allocated
Vec values of a Map.  This used to work by building a vector of the
keys, iterating over these keys, obtaining the values via get, and
deleting these values.  We now simply iterate over the Map using the
stylized loop form_Map.  This is significantly faster though the time
spent in this routine is relatively small.

For the compiler scaling test test_many_arrays_of_records, the time
spent in this routine decreases from 0.622s to 0.036s when compiling
with -sp=64 out of about 85s total time and from 2.653s to 0.079s when
compiling with -sp=128 out of about 255s total time.

Optimzed the part of reaching definitions analysis in which the defs
vector and defMap map are built such that the defs of the same
variable are adjacent in the vector and the map points from a def to a
location in this vector.  This used to use a map from Symbol pointers
to vectors of SymExpr pointers.  We now use an array of integers and a
map from symbols to integers (already computed).  This array is a
standard C array allocated over a variable size, i.e., something like
"int A[n];" where n is a variable of type int.  I don't believe we use
this idiom in the compiler much at all, but I believe it is C99.

For the compiler scaling test test_many_arrays_of_records, the time
spent in this part of the reaching definitions analysis decreases from
0.738s to 0.614s when compiling with -sp=64 out of about 85s total
time and from 3.378s to 2.238s when compiling with -sp=128 out of
about 255s total time.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16498 3a8e244f-b0f2-452b-bcba-4c88e055c3ca