Fixed a bug in which a call to a function could resolve incorrectly to a worse match that is more visible than a better match if there is an ambiguity between the better match and another function.  This could happen because we took a two-step approach to function resolution. First, call disambiguate_by_match to find the best function.  Then, if there is more than one possibility, call disambiguate_by_scope to eliminate all but the "most visible" functions as defined by distance between the call site and the function as measured by number of uses and number of parent scopes traversed.  Finally, call disambiguate_by_match a second time.
Removed a future.

Fixed the output of a point-of-instantiation test.  This test was
correct before.  This change to the output continues to be correct.
The interesting thing about this test is that there is a generic
function that has two potential instantiation points.  I know that one
is taken and the other is not.  Given which instantiation point is
taken, I expect the output to be as it is now.  However, if the other
instantiation point were taken, then I would expect it to be as it
was.  This test was thus showing a real bug, but hiding it due to the
multiple instantiation points.

The new tack for function resolution is to call a function
isMoreVisible with a call site and two functions.  It returns true if
the first function shadows the second function on all lexical and use
paths through the program from the specified call site.  For example,
I've annotated the following code to show which functions shadow
others with the same name and which do not.

  module M1 {
    def f() { }
    {
       def f() { } // shadows other f based on call to f
       f();
    }

    use M2;
    def g() { } // shadows other g based on call to g
    g();

    use M3;
    use M4;
    h();
  }

  module M2 { 
    def g() { }
  }

  module M3 {
    use M5;
    def h() { } // does not shadow other h based on call to h since
                // there is another path (through M4) to the other h
  }

  module M4 {
    use M5;
  }

  module M5 {
    def h() { }
  }

If a function does not shadow another function via all paths, then it
is not considered more visible.  If they are equally good matches, the
programmer would see an ambiguity.

Added two more tests of visibility disambiguation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16502 3a8e244f-b0f2-452b-bcba-4c88e055c3ca